requirements.txt
alpaca-trade-api>=3.0.0
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0
pytz>=2023.3
requests>=2.28.0
lxml>=4.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
